+++ nemo/export/huggingface/llama.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 from pathlib import Path
-from typing import Union
 
 import torch
 
@@ -32,7 +31,7 @@ from nemo.utils import logging
 LlamaModel = get_model("LlamaModel")
 
 
-# TODO: test multiple Llama model types
+# TODO: test multiple Llama model types, test Llama4
 @io_model_exporter(LlamaModel, "hf")
 class HFLlamaExporter(io.ModelConnector["LlamaModel", "LlamaForCausalLM"]):
     """Exporter for converting NeMo Llama models to Hugging Face format.
@@ -332,206 +331,3 @@ class HFLlamaExporter(io.ModelConnector["LlamaModel", "LlamaForCausalLM"]):
 
         return _ModelState(state_dict, source_config)
 
-
-# TODO: hf-peft exporter still utilises nemo.collections.llm
-# TODO: test before merging
-@io_model_exporter(LlamaModel, "hf-peft")
-class HFLlamaPEFTExporter(HFLlamaExporter):
-    """Exporter for converting NeMo Llama models with PEFT adapters to Hugging Face format.
-
-    This class extends HFLlamaExporter to handle Parameter-Efficient Fine-Tuning (PEFT)
-    adapters, specifically LoRA and DoRA adapters.
-    """
-
-    def init(self, dtype=torch.bfloat16) -> "AutoPeftModelForCausalLM":
-        """Initialize a HF PEFT model.
-
-        Args:
-            dtype: Data type for model parameters
-
-        Returns:
-            AutoPeftModelForCausalLM: Initialized HF PEFT model
-        """
-        import json
-
-        from peft import get_peft_model
-
-        from nemo.lightning.ckpt_utils import ADAPTER_META_FILENAME
-        from nemo.lightning.io.pl import ckpt_to_weights_subdir
-
-        model = super().init(dtype=dtype)
-
-        # Infer base model checkpoint from checkpoint metadata file
-        adapter_meta_path = ckpt_to_weights_subdir(str(self), is_saving=False) / ADAPTER_META_FILENAME
-        with open(adapter_meta_path, "r") as f:
-            model_ckpt_path = json.load(f)['model_ckpt_path']
-        model.name_or_path = '/'.join(model_ckpt_path.split("/")[-2:])
-
-        return get_peft_model(model, self.peft_config, autocast_adapter_dtype=False)
-
-    def apply(self, output_path: Path) -> Path:
-        """Apply the conversion from NeMo PEFT model to HF format.
-
-        Args:
-            output_path: Path where the converted model will be saved
-
-        Returns:
-            Path: Path to the saved HF PEFT model
-        """
-        from nemo.collections.llm.peft import CanonicalLoRA, DoRA, LoRA
-
-        self.peft_obj: Union[LoRA, DoRA, CanonicalLoRA] = io.load_context(str(self)).model.model_transform
-
-        source, _ = self.nemo_load(str(self))
-        target = self.init(torch_dtype_from_mcore_config(source.config))
-        target = self.convert_state(source, target)
-        target = target.cpu()
-        target.save_pretrained(output_path, save_embedding_layers=False)
-
-        return output_path
-
-    def convert_state(self, source, target):
-        """Convert state dict from NeMo PEFT model to HF PEFT format.
-
-        Maps the weights from the NeMo model to the HF model according to
-        the appropriate mapping scheme for PEFT adapters.
-
-        Args:
-            source: Source NeMo model with PEFT adapters
-            target: Target HF model
-
-        Returns:
-            The target model with weights transferred from source
-        """
-        from nemo.collections.llm.peft import CanonicalLoRA
-
-        # nemo and HF prefixes
-        pn = "decoder.layers."
-        ph = "base_model.model.model.layers."
-
-        # linear_proj and linear_fc2 prefixes
-        p_proj = "self_attention.linear_proj.adapter"
-        p_fc2 = "mlp.linear_fc2.adapter"
-
-        # linear_qkv and linear_fc1 prefixes
-        p_qkv = "self_attention.linear_qkv.adapter"
-        p_fc1 = "mlp.linear_fc1.adapter"
-
-        mapping = {
-            # linear_proj for both canonical and performant lora
-            f"{pn}*.{p_proj}.linear_in.weight": f"{ph}*.self_attn.o_proj.lora_A.default.weight",
-            f"{pn}*.{p_proj}.linear_out.weight": f"{ph}*.self_attn.o_proj.lora_B.default.weight",
-            # linear_fc2 for both canonical and performant lora
-            f"{pn}*.{p_fc2}.linear_in.weight": f"{ph}*.mlp.down_proj.lora_A.default.weight",
-            f"{pn}*.{p_fc2}.linear_out.weight": f"{ph}*.mlp.down_proj.lora_B.default.weight",
-        }
-        transforms = []
-
-        if isinstance(self.peft_obj, CanonicalLoRA):
-            mapping.update(
-                {
-                    # linear_qkv for canonical lora
-                    f"{pn}*.{p_qkv}.adapter_q.linear_in.weight": f"{ph}*.self_attn.q_proj.lora_A.default.weight",
-                    f"{pn}*.{p_qkv}.adapter_q.linear_out.weight": f"{ph}*.self_attn.q_proj.lora_B.default.weight",
-                    f"{pn}*.{p_qkv}.adapter_k.linear_in.weight": f"{ph}*.self_attn.k_proj.lora_A.default.weight",
-                    f"{pn}*.{p_qkv}.adapter_k.linear_out.weight": f"{ph}*.self_attn.k_proj.lora_B.default.weight",
-                    f"{pn}*.{p_qkv}.adapter_v.linear_in.weight": f"{ph}*.self_attn.v_proj.lora_A.default.weight",
-                    f"{pn}*.{p_qkv}.adapter_v.linear_out.weight": f"{ph}*.self_attn.v_proj.lora_B.default.weight",
-                    # linear_fc1 for canonical lora
-                    f"{pn}*.{p_fc1}.adapter_up.linear_in.weight": f"{ph}*.mlp.up_proj.lora_A.default.weight",
-                    f"{pn}*.{p_fc1}.adapter_up.linear_out.weight": f"{ph}*.mlp.up_proj.lora_B.default.weight",
-                    f"{pn}*.{p_fc1}.adapter_gate.linear_in.weight": f"{ph}*.mlp.gate_proj.lora_A.default.weight",
-                    f"{pn}*.{p_fc1}.adapter_gate.linear_out.weight": f"{ph}*.mlp.gate_proj.lora_B.default.weight",
-                }
-            )
-        else:
-            transforms.extend(
-                [
-                    # linear_qkv for performant lora
-                    io.state_transform(
-                        source_key=f"{pn}*.self_attention.linear_qkv.adapter.linear_in.weight",
-                        target_key=(
-                            f"{ph}*.self_attn.q_proj.lora_A.default.weight",
-                            f"{ph}*.self_attn.k_proj.lora_A.default.weight",
-                            f"{ph}*.self_attn.v_proj.lora_A.default.weight",
-                        ),
-                        fn=TransformFns.duplicate3,
-                    ),
-                    io.state_transform(
-                        source_key=f"{pn}*.self_attention.linear_qkv.adapter.linear_out.weight",
-                        target_key=(
-                            f"{ph}*.self_attn.q_proj.lora_B.default.weight",
-                            f"{ph}*.self_attn.k_proj.lora_B.default.weight",
-                            f"{ph}*.self_attn.v_proj.lora_B.default.weight",
-                        ),
-                        fn=TransformFns.split_qkv,
-                    ),
-                    # linear_fc1 for performant lora
-                    io.state_transform(
-                        source_key=f"{pn}*.mlp.linear_fc1.adapter.linear_in.weight",
-                        target_key=(
-                            f"{ph}*.mlp.gate_proj.lora_A.default.weight",
-                            f"{ph}*.mlp.up_proj.lora_A.default.weight",
-                        ),
-                        fn=TransformFns.duplicate2,
-                    ),
-                    io.state_transform(
-                        source_key=f"{pn}*.mlp.linear_fc1.adapter.linear_out.weight",
-                        target_key=(
-                            f"{ph}*.mlp.gate_proj.lora_B.default.weight",
-                            f"{ph}*.mlp.up_proj.lora_B.default.weight",
-                        ),
-                        fn=TransformFns.split_fc1,
-                    ),
-                ]
-            )
-
-        return io.apply_transforms(
-            source,
-            target,
-            mapping=mapping,
-            transforms=transforms,
-        )
-
-    @property
-    def peft_config(self) -> "PeftConfig":
-        """Create a PEFT config for the HF model.
-
-        Translates the NeMo PEFT configuration to the equivalent HF PEFT
-        configuration.
-
-        Returns:
-            PeftConfig: HF PEFT configuration
-        """
-        from peft import LoraConfig
-
-        from nemo.collections.llm.peft import DoRA
-
-        assert (
-            not self.peft_obj.dropout or self.peft_obj.dropout_position == 'pre'
-        ), "LoRA dropout_position must be 'pre' to convert to HF."
-
-        NEMO2HF = {
-            'linear_q': ['q_proj'],
-            'linear_k': ['k_proj'],
-            'linear_v': ['v_proj'],
-            'linear_qkv': ['q_proj', 'k_proj', 'v_proj'],
-            'linear_proj': ['o_proj'],
-            'linear_fc1_up': ['up_proj'],
-            'linear_fc1_gate': ['gate_proj'],
-            'linear_fc1': ['up_proj', 'gate_proj'],
-            'linear_fc2': ['down_proj'],
-        }
-
-        # Infer HF target modules from NeMo target modules
-        hf_target_modules = []
-        for tm in self.peft_obj.target_modules:
-            hf_target_modules.extend(NEMO2HF[tm])
-
-        return LoraConfig(
-            r=self.peft_obj.dim,
-            target_modules=hf_target_modules,
-            lora_alpha=self.peft_obj.alpha,
-            lora_dropout=self.peft_obj.dropout,
-            use_dora=isinstance(self.peft_obj, DoRA),
-        )
